\name{clusterId2SegmentIds}
\alias{clusterId2SegmentIds}
\title{clusterId2SegmentIds}
\description{Convert cluster IDs to indices of dendrogram segments.}
\usage{clusterId2SegmentIds(ids)}
\arguments{
  \item{ids}{cluster IDs}
}

\value{segment IDs}

\author{Tomas Sieger}






\keyword{internal}
