\name{computeLeafColorIdxs}
\alias{computeLeafColorIdxs}
\title{computeLeafColorIdxs}
\description{Compute colors of all leafs in dendrogram according to the current
cluster selection.}
\usage{computeLeafColorIdxs(df)}
\arguments{
  \item{df}{dhared data frame, see 'prepareDendro'}
}

\value{leaf colors}

\author{Tomas Sieger}






\keyword{internal}
