\name{computeLeafOrder}
\alias{computeLeafOrder}
\title{computeLeafOrder}
\description{Compute the assignment of dendrogram leafs to the original
observations.}
\usage{computeLeafOrder(merging, dbg = 0)}
\arguments{
  \item{merging}{(n-1) by 2 matrix describing HCA merging, usually the
'merge' component of the return value from 'hclust'}
  \item{dbg}{debug verbosity level}
}

\value{leaf ordering}

\author{Tomas Sieger}






\keyword{internal}
