\name{computeSubclusterIndices}
\alias{computeSubclusterIndices}
\title{computeSubclusterIndices}

\usage{computeSubclusterIndices(h, clusterIdx)}
\arguments{
  \item{h}{HCA result, usually 'hclust' return value}
  \item{clusterIdx}{cluster index}
}

\value{subclusters indices}

\author{Tomas Sieger}





