\name{dendro2fig}
\alias{dendro2fig}
\title{dendro2fig}
\description{Dendrogram -> figure coords conversion.}
\usage{dendro2fig(gw)}
\arguments{
  \item{gw}{a list of 'g' and 'w' components, g=gw[1]: dendro height ... 0,
w=gw[2]: n ... 1 (charm to strange)}
}

\value{converted values, a list of 'g' and 'w' components}

\author{Tomas Sieger}






\keyword{internal}
