\name{fig2dendro}
\alias{fig2dendro}
\title{fig2dendro}
\description{Fig -> dendro coord conversion.}
\usage{fig2dendro(gw)}
\arguments{
  \item{gw}{##<< a list of 'g' and 'w' components}
}

\value{converted values, a list of 'g' and 'w' components}

\author{Tomas Sieger}






\keyword{internal}
