\name{fig2heatmap}
\alias{fig2heatmap}
\title{fig2heatmap}
\description{Figure -> heatmap coord conversion.}
\usage{fig2heatmap(gw)}
\arguments{
  \item{gw}{a list of 'g' and 'w' components}
}

\value{a list of 'g' and 'w' components}

\author{Tomas Sieger}






\keyword{internal}
