\name{heatmap2fig}
\alias{heatmap2fig}
\title{heatmap2fig}
\description{Heatmap -> figure coord conversion.}
\usage{heatmap2fig(gw)}
\arguments{
  \item{gw}{a list of 'g' and 'w' components}
}

\value{a list of 'g' and 'w' components}

\author{Tomas Sieger}






\keyword{internal}
