\name{popDendroZoomHistory}
\alias{popDendroZoomHistory}
\title{popDendroZoomHistory}
\description{Restore (and discard from history) the last dendro zoom.}
\usage{popDendroZoomHistory(.sharedEnv)}
\arguments{
  \item{.sharedEnv}{shared environment (the environment of the
'idendro' function)}
}

\value{dendro zoom}

\author{Tomas Sieger}






\keyword{internal}
