\name{pushDendroZoomHistory}
\alias{pushDendroZoomHistory}
\title{pushDendroZoomHistory}
\description{Save the current dendroZoom (as stored in '.sharedEnv').}
\usage{pushDendroZoomHistory(.sharedEnv)}
\arguments{
  \item{.sharedEnv}{shared environment (the environment of the
'idendro' function)}
}



\author{Tomas Sieger}






\keyword{internal}
