\name{selectCluster}
\alias{selectCluster}
\title{selectCluster}
\description{Select the nearest cluster to the position 'pos'.}
\usage{selectCluster(pos)}
\arguments{
  \item{pos}{position in the dendro figure}
}

\value{shared data frame}

\author{Tomas Sieger}






\keyword{internal}
