\name{smoothHeatmapAccordingToClusters}
\alias{smoothHeatmapAccordingToClusters}
\title{smoothHeatmapAccordingToClusters}
\description{Smooth heatmap by averaging the observations associated with the
currently selected clusters. Observations that are not contained
in any cluster currently selected stay untouched.}
\usage{smoothHeatmapAccordingToClusters(df, dbg.heatmap.smooth = 0)}
\arguments{
  \item{df}{shared data frame}
  \item{dbg.heatmap.smooth}{debug verbosity level}
}



\author{Tomas Sieger}






\keyword{internal}
