\name{computeMemberIndices}
\alias{computeMemberIndices}
\title{computeMemberIndices}
\description{Compute indices of observations forming given cluster.}
\usage{computeMemberIndices(h, clusterIdx)}
\arguments{
  \item{h}{HCA result, usually 'hclust' return value}
  \item{clusterIdx}{cluster index}
}

\value{indices of observations forming cluster 'clusterIdx'}

\author{Tomas Sieger}






\keyword{internal}
