\name{computeSubclusterIndices}
\alias{computeSubclusterIndices}
\title{computeSubclusterIndices}
\description{Compute indices of all subclusters (even deep down the hierarchy)
appearing in given cluster.}
\usage{computeSubclusterIndices(h, clusterIdx)}
\arguments{
  \item{h}{HCA result, usually 'hclust' return value}
  \item{clusterIdx}{cluster index}
}

\value{subclusters indices}

\author{Tomas Sieger}






\keyword{internal}
