\name{idendro-package}
\alias{idendro-package}
\docType{package}
\title{Interactive dendrogram}
\description{Interactive dendrogram (with optional heatmap) enabling
cluster selection, zooming, panning and bidirectional integration
with other plots and tools by using mutable data frames.}
\details{
\tabular{ll}{Package: \tab idendro\cr
Type: \tab Package\cr
Title: \tab Interactive dendrogram\cr
Version: \tab 1.2.0\cr
Date: \tab 2012-11-18\cr
Author: \tab Tomas Sieger\cr
Depends: \tab qtbase, qtpaint, cranvas, plumbr, grDevices, scales\cr
URL: \tab http://github.com/tsieger/idendro\cr
Maintainer: \tab Tomas Sieger <tomas.sieger@seznam.cz>\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr}
}
\author{Tomas Sieger}

\keyword{ package }


