\name{idendro-package}
\alias{idendro-package}
\docType{package}
\title{Interactive dendrograms}
\description{Interactive dendrogram that enables the user to select and
color clusters, to zoom and pan the dendrogram, and to visualize
the clustered data not only in a built-in heat map, but also in any
interactive plot implemented in the 'cranvas' package.}
\details{
\tabular{ll}{Package: \tab idendro\cr
Type: \tab Package\cr
Title: \tab Interactive dendrograms\cr
Version: \tab 1.5.2.9000\cr
Date: \tab 2015-10-14\cr
Author: \tab Tomas Sieger\cr
Depends: \tab cranvas, plumbr, grDevices\cr
Imports: \tab qtbase, qtpaint, scales\cr
Suggests: \tab DendSer, cluster, tourr, RColorBrewer, testthat\cr
URL: \tab http://github.com/tsieger/idendro\cr
Maintainer: \tab Tomas Sieger <tomas.sieger@seznam.cz>\cr
BugReports: \tab https://github.com/tsieger/idendro/issues\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr}
}
\author{Tomas Sieger}

\keyword{ package }

\references{
  Sieger, T.,  Hurley, C. B., Fiser, K., Beleites, C. (2017)
  Interactive Dendrograms: The R Packages idendro and idendr0.
  \emph{Journal of Statistical Software}, \bold{76}(10), 1--22.
  doi:10.18637/jss.v076.i10
}
