\name{idendro-package}
\alias{idendro-package}
\alias{idendro}
\docType{package}
\title{Interactive dendrograms}
\description{Interactive dendrogram that enables the user to select and
color clusters, to zoom and pan the dendrogram, and to visualize
the clustered data not only in a built-in heat map, but also in any
interactive plot implemented in the cranvas package.}
\details{
\tabular{ll}{Package: \tab idendro\cr
Type: \tab Package\cr
Title: \tab Interactive dendrograms\cr
Version: \tab 1.4.2\cr
Date: \tab 2015-04-10\cr
Author: \tab Tomas Sieger\cr
Depends: \tab qtbase, qtpaint, cranvas, plumbr, grDevices, scales\cr
Suggests: \tab DendSer, testthat\cr
URL: \tab http://github.com/tsieger/idendro\cr
Maintainer: \tab Tomas Sieger <tomas.sieger@seznam.cz>\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr}
}
\author{Tomas Sieger}

\keyword{ package }


