\name{idendro-package}
\alias{idendro-package}
\docType{package}
\title{Interactive dendrogram}
\description{Interactive dendrogram + heat map plot enabling
cluster selection, zooming, panning and integration
with other interactive plots.}
\details{
\tabular{ll}{Package: \tab idendro\cr
Type: \tab Package\cr
Title: \tab Interactive dendrogram\cr
Version: \tab 1.2.1\cr
Date: \tab 2012-11-18\cr
Author: \tab Tomas Sieger\cr
Depends: \tab qtbase, qtpaint, cranvas, plumbr, grDevices, scales\cr
Suggests: \tab DendSer\cr
URL: \tab http://github.com/tsieger/idendro\cr
Maintainer: \tab Tomas Sieger <tomas.sieger@seznam.cz>\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr}
}
\author{Tomas Sieger}

\keyword{ package }


