\name{popDendroZoomHistory}
\alias{popDendroZoomHistory}
\title{popDendroZoomHistory}
\description{Restore (and discard from history) the last dendro zoom.}
\usage{popDendroZoomHistory(df, dbg = FALSE)}
\arguments{
  \item{df}{shared data frame}
  \item{dbg}{debug flag/level}
}

\value{dendro zoom popped (or NULL if stack empty)}

\author{Tomas Sieger}






\keyword{internal}
