\name{popSelectionHistory}
\alias{popSelectionHistory}
\title{popSelectionHistory}
\description{Restore (and discard from history) the last cluster selection.}
\usage{popSelectionHistory(df, dbg = FALSE)}
\arguments{
  \item{df}{shared data frame}
  \item{dbg}{debug flag/level}
}

\value{shared data frame holding cluster selection (or NULL if no
selection found in the selection history)}

\author{Tomas Sieger}






\keyword{internal}
