\name{pushDendroZoomHistory}
\alias{pushDendroZoomHistory}
\title{pushDendroZoomHistory}
\description{Save the current dendroZoom (as stored in '.sharedEnv').}
\usage{pushDendroZoomHistory(df, dendroZoom, dbg = FALSE)}
\arguments{
  \item{df}{shared data frame}
  \item{dendroZoom}{dendro zoom to push}
  \item{dbg}{debug flag/level}
}

\value{shared data frame with dendro zoom pushed on top of the
dendro zoom stack}

\author{Tomas Sieger}






\keyword{internal}
