\name{pushSelectionHistory}
\alias{pushSelectionHistory}
\title{pushSelectionHistory}
\description{Save the current cluster selection (as stored in df).}
\usage{pushSelectionHistory(df, dbg = FALSE)}
\arguments{
  \item{df}{shared data frame}
  \item{dbg}{debug flag/level}
}

\value{shared data frame with cluster selection pushed on top of the
selection stack.}

\author{Tomas Sieger}






\keyword{internal}
