\name{selectCluster}
\alias{selectCluster}
\title{selectCluster}
\description{Select the nearest cluster to the position 'pos'.}
\usage{selectCluster(pos, df, dendroZoom, dbg = FALSE)}
\arguments{
  \item{pos}{(x,y) position in the dendro figure}
  \item{df}{shared data frame}
  \item{dendroZoom}{the current dendro zoom region (used to determine
the scaling)}
  \item{dbg}{debug flag/level}
}

\value{shared data frame}

\author{Tomas Sieger}






\keyword{internal}
